+++ CRM/Civirules/Trigger.php
@@ -51,6 +51,17 @@ abstract class CRM_Civirules_Trigger {
    */
   abstract protected function reactOnEntity();
 
+  /**
+   * Returns the name of the trigger data class.
+   *
+   * This function could be overridden in a child class.
+   *
+   * @return String
+   */
+  public function getTriggerDataClassName() {
+    return 'CRM_Civirules_TriggerData_TriggerData';
+  }
+
 
   public function getProvidedEntities() {
     $additionalEntities = $this->getAdditionalEntities();
@@ -62,6 +73,13 @@ abstract class CRM_Civirules_Trigger {
     $entities[$entity->key] = $entity;
 
     return $entities;
+  }
+
+  /**
+   * @return \CRM_Civirules_TriggerData_EntityDefinition
+   */
+  public function getReactOnEntity() {
+    return $this->reactOnEntity();
   }
 
   /**