+++ CRM/CivirulesConditions/Form/FieldValueComparison.php
@@ -6,6 +6,8 @@
  * @license AGPL-3.0
  */
 
+use CRM_Civirules_ExtensionUtil as E;
+
 class CRM_CivirulesConditions_Form_FieldValueComparison extends CRM_CivirulesConditions_Form_ValueComparison {
 
   protected function getEntityOptions() {
@@ -108,8 +110,8 @@ class CRM_CivirulesConditions_Form_FieldValueComparison extends CRM_CivirulesCon
     parent::buildQuickForm();
 
     $this->add('hidden', 'rule_condition_id');
-    $this->add('select', 'entity', ts('Entity'), $this->getEntityOptions(), true);
-    $this->add('select', 'field', ts('Field'), $this->getFields(), true, array('class' => 'crm-select2'));
+    $this->add('select', 'entity', ts('Entity'), $this->getEntityOptions(), true, array('class' => 'crm-select2 huge'));
+    $this->add('select', 'field', ts('Field'), $this->getFields(), true, array('class' => 'crm-select2 huge'));
     $this->add('checkbox', 'original_data', ts('Compare with original value (before the change)?'));
     $this->assign('entities', $this->getEntities());
     $this->assign('custom_field_multi_select_html_types', CRM_Civirules_Utils_CustomField::getMultiselectTypes());
@@ -153,6 +155,7 @@ class CRM_CivirulesConditions_Form_FieldValueComparison extends CRM_CivirulesCon
     if (!empty($data['entity']) && !empty($data['field'])) {
       $defaultValues['field'] = $data['entity'].'_'.$data['field'];
     }
+    $defaultValues['original_data'] = $data['original_data'] ?? 0;
     return $defaultValues;
   }
 
@@ -179,9 +182,10 @@ class CRM_CivirulesConditions_Form_FieldValueComparison extends CRM_CivirulesCon
     $this->ruleCondition->save();
 
     $session = CRM_Core_Session::singleton();
-    $session->setStatus('Condition '.$this->condition->label .'Parameters updated to CiviRule '
-      .$this->rule->label,
-      'Condition parameters updated', 'success');
+    $session->setStatus(E::ts("Condition '%1' parameters updated for CiviRule '%2'", [1 => $this->condition->label, 2 => $this->rule->label]),
+      'Condition parameters updated',
+      'success'
+    );
 
     $redirectUrl = CRM_Utils_System::url('civicrm/civirule/form/rule', 'action=update&id='.$this->rule->id, TRUE);
     CRM_Utils_System::redirect($redirectUrl);  }
@@ -193,7 +197,9 @@ class CRM_CivirulesConditions_Form_FieldValueComparison extends CRM_CivirulesCon
    */
   protected function setFormTitle() {
     $title = 'CiviRules Edit Condition parameters';
-    $this->assign('ruleConditionHeader', 'Edit Condition '.$this->condition->label.' of CiviRule '.$this->rule->label);
+    $this->assign('ruleConditionHeader',
+      E::ts("Edit Condition '%1' for CiviRule '%2'", [ 1 => $this->condition->label, 2 => $this->rule->label])
+    );
     CRM_Utils_System::setTitle($title);
   }
 }