+++ CRM/CivirulesConditions/Membership/Status.php
@@ -1,102 +1,48 @@
 <?php
 
-class CRM_CivirulesConditions_Membership_Status extends CRM_Civirules_Condition {
-
-  private $conditionParams = array();
+class CRM_CivirulesConditions_Membership_Status extends CRM_CivirulesConditions_Generic_Status {
 
   /**
-   * Method to set the Rule Condition data
-   *
-   * @param array $ruleCondition
-   * @access public
-   */
-  public function setRuleConditionData($ruleCondition) {
-    parent::setRuleConditionData($ruleCondition);
-    $this->conditionParams = array();
-    if (!empty($this->ruleCondition['condition_params'])) {
-      $this->conditionParams = unserialize($this->ruleCondition['condition_params']);
-    }
-  }
-
-  /**
-   * Method to determine if the condition is valid
-   *
-   * @param CRM_Civirules_TriggerData_TriggerData $triggerData
-   * @return bool
-   */
-  public function isConditionValid(CRM_Civirules_TriggerData_TriggerData $triggerData) {
-    $isConditionValid = FALSE;
-    $membership = $triggerData->getEntityData('Membership');
-    switch ($this->conditionParams['operator']) {
-      case 0:
-        if ($membership['status_id'] == $this->conditionParams['membership_status_id']) {
-          $isConditionValid = TRUE;
-        }
-      break;
-      case 1:
-        if ($membership['status_id'] != $this->conditionParams['membership_status_id']) {
-          $isConditionValid = TRUE;
-        }
-      break;
-    }
-    return $isConditionValid;
-  }
-
-  /**
-   * Returns a redirect url to extra data input from the user after adding a condition
-   *
-   * Return false if you do not need extra data input
-   *
-   * @param int $ruleConditionId
-   * @return bool|string
-   * @access public
-   * @abstract
-   */
-  public function getExtraDataInputUrl($ruleConditionId) {
-    return CRM_Utils_System::url('civicrm/civirule/form/condition/membershipstatus', 'rule_condition_id='
-      .$ruleConditionId);
-  }
-
-  /**
-   * Returns a user friendly text explaining the condition params
-   * e.g. 'Older than 65'
-   *
+   * The entity name (eg. Membership)
    * @return string
-   * @access public
    */
-  public function userFriendlyConditionParams() {
+  protected function getEntity() {
+    return 'Membership';
+  }
+
+  /**
+   * The entity status field (eg. membership_status_id)
+   * @return string
+   */
+  public function getEntityStatusFieldName() {
+    return 'status_id';
+  }
+
+  /**
+   * Returns an array of statuses as [ id => label ]
+   * @param bool $active
+   * @param bool $inactive
+   *
+   * @return array
+   */
+  public static function getEntityStatusList($active = TRUE, $inactive = FALSE) {
+    $return = [];
+    $params = [];
+    if ($active && !$inactive) {
+      $params = ['is_active' => 1];
+    }
+    elseif ($inactive && !$active) {
+      $params = ['is_active' => 0];
+    }
+    $params['options'] = ['limit' => 0, 'sort' => "label ASC"];
+
     try {
-      $apiMembershipStatus = civicrm_api3('MembershipStatus', 'Get', array('is_active' => 1));
-      $operator = null;
-      if ($this->conditionParams['operator'] == 0) {
-        $operator = 'equals';
-      }
-      if ($this->conditionParams['operator'] == 1) {
-        $operator = 'is not equal to';
-      }
-      foreach ($apiMembershipStatus['values'] as $membershipStatus) {
-        if ($membershipStatus['id'] == $this->conditionParams['membership_status_id']) {
-          return "Membership Status ".$operator." ".$membershipStatus['name'];
-        }
+      $apiMembershipStatus = civicrm_api3("MembershipStatus", "Get", $params)['values'];
+      foreach ($apiMembershipStatus as $membershipStatus) {
+        $return[$membershipStatus['id']] = $membershipStatus['label'];
       }
     } catch (CiviCRM_API3_Exception $ex) {}
-    return '';
-  }
-
-  /**
-   * This function validates whether this condition works with the selected trigger.
-   *
-   * This function could be overriden in child classes to provide additional validation
-   * whether a condition is possible in the current setup. E.g. we could have a condition
-   * which works on contribution or on contributionRecur then this function could do
-   * this kind of validation and return false/true
-   *
-   * @param CRM_Civirules_Trigger $trigger
-   * @param CRM_Civirules_BAO_Rule $rule
-   * @return bool
-   */
-  public function doesWorkWithTrigger(CRM_Civirules_Trigger $trigger, CRM_Civirules_BAO_Rule $rule) {
-    return $trigger->doesProvideEntity('Membership');
+    return $return;
   }
 
 }