+++ CRM/CivirulesCronTrigger/ActivityDate.php
@@ -4,56 +4,14 @@
  * @license http://www.gnu.org/licenses/agpl-3.0.html
  */
 
-class CRM_CivirulesCronTrigger_ActivityDate extends CRM_Civirules_Trigger_Cron {
+use CRM_Civirules_ExtensionUtil as E;
 
-  private $dao = false;
-
-  /**
-   * This function returns a CRM_Civirules_TriggerData_TriggerData this entity is used for triggering the rule
-   *
-   * Return false when no next entity is available
-   *
-   * @return CRM_Civirules_TriggerData_TriggerData|false
-   */
-  protected function getNextEntityTriggerData() {
-    if (!$this->dao) {
-      if (!$this->queryForTriggerEntities()) {
-        return false;
-      }
-    }
-    if ($this->dao->fetch()) {
-      $data = array();
-      CRM_Core_DAO::storeValues($this->dao, $data);
-      unset($data['activity_contact_id']);
-      unset($data['contact_id']);
-      unset($data['record_type_id']);
-      $triggerData = new CRM_Civirules_TriggerData_Cron($this->dao->contact_id, 'Activity', $data);
-      $activityContact = array();
-      $activityContact['id'] = $this->dao->activity_contact_id;
-      $activityContact['activity_id'] = $this->dao->id;
-      $activityContact['contact_id'] = $this->dao->contact_id;
-      $activityContact['record_type_id'] = $this->dao->record_type_id;
-      $triggerData->setEntityData('ActivityContact', $activityContact);
-      return $triggerData;
-    }
-    return false;
-  }
-
-  /**
-   * Returns an array of entities on which the trigger reacts
-   *
-   * @return CRM_Civirules_TriggerData_EntityDefinition
-   */
-  protected function reactOnEntity() {
-    return new CRM_Civirules_TriggerData_EntityDefinition('Activity', 'Activity', 'CRM_Activity_DAO_Activity', 'Activity');
-  }
+class CRM_CivirulesCronTrigger_ActivityDate extends CRM_CivirulesCronTrigger_Activity {
 
   /**
    * Method to query trigger entities
-   *
-   * @access private
    */
-  private function queryForTriggerEntities() {
+  protected function queryForTriggerEntities() {
     if (empty($this->triggerParams['activity_type_id'])) {
       return false;
     }
@@ -61,19 +19,35 @@ class CRM_CivirulesCronTrigger_ActivityDate extends CRM_Civirules_Trigger_Cron {
       return false;
     }
 
-    $sql = "SELECT a.*, ac.contact_id as contact_id, ac.record_type_id as record_type_id, ac.id as activity_contact_id
+    $activityContactWhereClause = '';
+    if (!empty($this->triggerParams['record_type'])) {
+      $activityContactWhereClause = "AND `ac`.`record_type_id` = %5";
+      $params[5] = [$this->triggerParams['record_type'], 'Integer'];
+    }
+
+    $activityCaseWhereClause = 'AND `ca`.`case_id` IS NULL';
+    if (!empty($this->triggerParams['case_activity'])) {
+      $activityCaseWhereClause = 'AND `ca`.`case_id` IS NOT NULL';
+    }
+
+    $sql = "SELECT a.*, ac.contact_id as contact_id, ac.record_type_id as record_type_id, ac.id as activity_contact_id, ca.case_id as case_id
             FROM `civicrm_activity` `a`
             INNER JOIN `civicrm_activity_contact` ac ON a.id = ac.activity_id
+            LEFT JOIN `civicrm_case_activity` ca ON a.id = ca.activity_id
+            LEFT JOIN `civirule_rule_log` `rule_log` ON `rule_log`.entity_table = 'civicrm_activity' AND `rule_log`.entity_id = a.id AND `rule_log`.`contact_id` = `ac`.`contact_id` AND DATE(`rule_log`.`log_date`) = DATE(NOW())  AND `rule_log`.`rule_id` = %3
             WHERE `a`.`activity_type_id` = %1 AND a.status_id = %2 AND a.activity_date_time <= NOW()
+            AND `rule_log`.`id` IS NULL
+            {$activityContactWhereClause}
+            {$activityCaseWhereClause}
             AND `ac`.`contact_id` NOT IN (
-              SELECT `rule_log`.`contact_id`
-              FROM `civirule_rule_log` `rule_log`
-              WHERE `rule_log`.`rule_id` = %3 AND DATE(`rule_log`.`log_date`) = DATE(NOW())
+              SELECT `rule_log2`.`contact_id`
+              FROM `civirule_rule_log` `rule_log2`
+              WHERE `rule_log2`.`rule_id` = %3 AND DATE(`rule_log2`.`log_date`) = DATE(NOW()) and `rule_log2`.`entity_table` IS NULL AND `rule_log2`.`entity_id` IS NULL
             )";
     $params[1] = array($this->triggerParams['activity_type_id'], 'Integer');
     $params[2] = array($this->triggerParams['activity_status_id'], 'Integer');
     $params[3] = array($this->ruleId, 'Integer');
-    $this->dao = CRM_Core_DAO::executeQuery($sql, $params, true, 'CRM_Activity_DAO_Activity');
+    $this->activityDAO = CRM_Core_DAO::executeQuery($sql, $params, true, 'CRM_Activity_DAO_Activity');
 
     return true;
   }
@@ -84,41 +58,39 @@ class CRM_CivirulesCronTrigger_ActivityDate extends CRM_Civirules_Trigger_Cron {
    * Return false if you do not need extra data input
    *
    * @param int $ruleId
+   *
    * @return bool|string
-   * @access public
-   * @abstract
    */
   public function getExtraDataInputUrl($ruleId) {
-    return CRM_Utils_System::url('civicrm/civirule/form/trigger/activitydate/', 'rule_id='.$ruleId);
-  }
-
-  public function setTriggerParams($triggerParams) {
-    $this->triggerParams = unserialize($triggerParams);
+    return CRM_Utils_System::url('civicrm/civirule/form/trigger/activitydate', 'rule_id='.$ruleId);
   }
 
   /**
    * Returns a description of this trigger
    *
    * @return string
-   * @access public
-   * @abstract
    */
   public function getTriggerDescription() {
     $activityTypeLabel = CRM_Civirules_Utils::getOptionLabelWithValue(CRM_Civirules_Utils::getOptionGroupIdWithName('activity_type'),  $this->triggerParams['activity_type_id']);
     $activityStatusLabel = CRM_Civirules_Utils::getOptionLabelWithValue(CRM_Civirules_Utils::getOptionGroupIdWithName('activity_status'),  $this->triggerParams['activity_status_id']);
 
-    return ts('Activity with type %1 and status %2 date reached', array(1 => $activityTypeLabel, 2=> $activityStatusLabel));
-  }
+    $result = civicrm_api3('ActivityContact', 'getoptions', [
+      'field' => "record_type_id",
+    ]);
+    $options[0] = E::ts('All contacts');
+    $options = array_merge($options, $result['values']);
 
-  /**
-   * Returns an array of additional entities provided in this trigger
-   *
-   * @return array of CRM_Civirules_TriggerData_EntityDefinition
-   */
-  protected function getAdditionalEntities() {
-    $entities = parent::getAdditionalEntities();
-    $entities[] = new CRM_Civirules_TriggerData_EntityDefinition('ActivityContact', 'ActivityContact', 'CRM_Activity_DAO_ActivityContact' , 'ActivityContact');
-    return $entities;
+    $caseActivity = 'Not case activity';
+    if (!empty($this->triggerParams['case_activity'])) {
+      $caseActivity = 'Case activity';
+    }
+
+    return ts('%4 with type %1 and status %2 date reached. Trigger for %3', [
+      1 => $activityTypeLabel,
+      2 => $activityStatusLabel,
+      3 => $options[$this->triggerParams['record_type']],
+      4 => $caseActivity,
+    ]);
   }
 
 }