+++ CRM/CivirulesCronTrigger/Form/ActivityDate.php
@@ -3,47 +3,24 @@
  * @author Jaap Jansma (CiviCooP) <jaap.jansma@civicoop.org>
  * @license http://www.gnu.org/licenses/agpl-3.0.html
  */
-class CRM_CivirulesCronTrigger_Form_ActivityDate extends CRM_CivirulesTrigger_Form_Form {
 
-  protected function getActivityType() {
-    return CRM_Civirules_Utils::getActivityTypeList();
-  }
+use CRM_Civirules_ExtensionUtil as E;
 
-  protected function getActivityStatus() {
-    $activityStatusList = array();
-    $activityStatusOptionGroupId = CRM_Civirules_Utils::getOptionGroupIdWithName('activity_status');
-    $params = array(
-      'option_group_id' => $activityStatusOptionGroupId,
-      'is_active' => 1,
-      'options' => array('limit' => 0));
-    $activityStatuses = civicrm_api3('OptionValue', 'Get', $params);
-    foreach ($activityStatuses['values'] as $optionValue) {
-      $activityStatusList[$optionValue['value']] = $optionValue['label'];
-    }
-    return $activityStatusList;
-  }
+class CRM_CivirulesCronTrigger_Form_ActivityDate extends CRM_CivirulesCronTrigger_Form_Activity {
 
   /**
    * Overridden parent method to build form
-   *
-   * @access public
    */
   public function buildQuickForm() {
-    $this->add('hidden', 'rule_id');
-
-    $this->add('select', 'activity_type_id', ts('Activity Type'), $this->getActivityType(), true);
-    $this->add('select', 'activity_status_id', ts('Activity Status'), $this->getActivityStatus(), true);
-
-    $this->addButtons(array(
-      array('type' => 'next', 'name' => ts('Save'), 'isDefault' => TRUE,),
-      array('type' => 'cancel', 'name' => ts('Cancel'))));
+    parent::buildQuickForm();
+    $this->add('select', 'activity_type_id', ts('Activity Type'), $this->getActivityType(), TRUE);
+    $this->add('select', 'activity_status_id', ts('Activity Status'), $this->getActivityStatus(), TRUE);
   }
 
   /**
    * Overridden parent method to set default values
    *
    * @return array $defaultValues
-   * @access public
    */
   public function setDefaultValues() {
     $defaultValues = parent::setDefaultValues();
@@ -54,6 +31,15 @@ class CRM_CivirulesCronTrigger_Form_ActivityDate extends CRM_CivirulesTrigger_Fo
     if (!empty($data['activity_status_id'])) {
       $defaultValues['activity_status_id'] = $data['activity_status_id'];
     }
+
+    if (!empty($data['record_type'])) {
+      $defaultValues['record_type'] = $data['record_type'];
+    } else {
+      $defaultValues['record_type'] = 3; // Default to only targets
+    }
+
+    $defaultValues['case_activity'] = $data['case_activity'] ?? FALSE;
+
     return $defaultValues;
   }
 
@@ -61,14 +47,28 @@ class CRM_CivirulesCronTrigger_Form_ActivityDate extends CRM_CivirulesTrigger_Fo
    * Overridden parent method to process form data after submission
    *
    * @throws Exception when rule condition not found
-   * @access public
    */
   public function postProcess() {
     $data['activity_type_id'] = $this->_submitValues['activity_type_id'];
     $data['activity_status_id'] = $this->_submitValues['activity_status_id'];
+    $data['record_type'] = $this->_submitValues['record_type'];
+    $data['case_activity'] = $this->_submitValues['case_activity'];
     $this->rule->trigger_params = serialize($data);
     $this->rule->save();
 
     parent::postProcess();
   }
+
+  /**
+   * Returns a help text for this trigger.
+   * The help text is shown to the administrator who is configuring the condition.
+   *
+   * @return string
+   */
+  protected function getHelpText() {
+    return E::ts('Trigger rule when scheduled date for activity with status and type is reached.')
+      . '<br/>'
+      . E::ts('If "Trigger for case activities" is "Yes" then this will only trigger for case activities. If it is "No" then it will only trigger for activities that are not linked to a case.');
+  }
+
 }